Исправил Enums и изменил структуру
