+++ src/main/java/api/dto/ProductDTO.java
@@ -1,4 +1,4 @@
-package api.version_1.dto;
+package api.dto;
 
 import lombok.Data;
