+++ src/main/java/api/enumclass/OrderStatus.java
@@ -1,0 +1,8 @@
+package api.enumclass;
+
+public enum OrderStatus {
+    created,
+    processing,
+    cancelled,
+    shipped
+}
