+++ src/main/java/api/enumclass/PaymentMethod.java
@@ -1,0 +1,7 @@
+package api.enumclass;
+
+public enum PaymentMethod {
+    CREDIT_CARD,
+    PAY_PAL,
+    ON_RECEIPT
+}
