+++ src/main/java/api/repository/OrderDetailRepository.java
@@ -1,6 +1,6 @@
-package api.version_1.repository;
+package api.repository;
 
-import api.version_1.entity.OrderDetail;
+import api.entity.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
