+++ src/main/java/api/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package api.version_1.repository;
+package api.repository;
 
-import api.version_1.entity.Product;
+import api.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
