--- src/main/java/api/version_1/Enums.java
@@ -1,16 +1,0 @@
-package api.version_1;
-
-public class Enums {
-
-    public enum OrderStatus {
-        created,
-        processing,
-        cancelled,
-        shipped
-    };
-    public enum PaymentMethod {
-        CREDIT_CARD,
-        PAY_PAL,
-        ON_RECEIPT
-    }
-}
